Refactor monolithic architecture into component-based structure
This commit addresses the user's request for "errors and improvements" by performing a major refactoring of the application's architecture. The previous implementation consisted of a single, 2000+ line `index.tsx` file, which was difficult to maintain, debug, and scale.

The main changes include:

- **Componentization:** The monolithic `index.tsx` file has been broken down into a standard React component structure. All views, shared components, and modals have been extracted into their own files under `src/components`.

- **Data Separation:** Static data objects (e.g., `healthAreas`, `stretchingExercises`) have been moved from the main component file into a dedicated `src/data` directory, making them easier to manage.

- **Icon Abstraction:** All inline SVG icon components have been moved to a single `src/icons.tsx` file for better organization and reusability.

- **Security Improvement:** The hardcoded Gemini API key access was changed to use Vite's environment variables (`import.meta.env.VITE_GEMINI_API_KEY`). A security warning was added to the code to advise against exposing the key on the client side in a production environment.

- **Robustness:** Added a `try-catch` block for `localStorage` parsing to prevent the application from crashing due to corrupted user data.

- **Bug Fixes:** Removed several unused code blocks and comments that were marked with `FIXME`.

This refactoring does not change the application's functionality but significantly improves its code quality, maintainability, and scalability, while also addressing a key security vulnerability and a potential runtime error.
