+++ src/data/focusMusicTracks.ts
@@ -1,0 +1,9 @@
+import { CloudRainIcon, FlameIcon, WavesIcon, TreePineIcon, MusicIcon } from '../icons';
+
+export const focusMusicTracks = [
+    { name: 'Chuva Leve', icon: CloudRainIcon, url: 'https://cdn.pixabay.com/audio/2022/10/21/audio_18c593256a.mp3' },
+    { name: 'Fogueira Aconchegante', icon: FlameIcon, url: 'https://cdn.pixabay.com/audio/2022/11/15/audio_1380907106.mp3' },
+    { name: 'Ondas do Mar', icon: WavesIcon, url: 'https://cdn.pixabay.com/audio/2023/10/18/audio_2910712739.mp3' },
+    { name: 'Floresta Tranquila', icon: TreePineIcon, url: 'https://cdn.pixabay.com/audio/2023/08/03/audio_5119154f24.mp3' },
+    { name: 'Música Ambiente Calma', icon: MusicIcon, url: 'https://cdn.pixabay.com/audio/2022/08/04/audio_2dde648d15.mp3' }
+];
