+++ src/data/motivationalPhrases.ts
@@ -1,0 +1,1 @@
+export const generalMotivationalPhrases = ['Pequenos passos levam a grandes jornadas!', 'Hoje melhor que ontem, amanhã melhor que hoje!', 'A virtude está na ação, não na intenção!', 'Cada escolha é uma oportunidade de crescer!', 'O Kaizen transforma sonhos em realidade!', 'Sua jornada extraordinária começa agora!', 'A serenidade vem da sabedoria em ação!', 'Você tem o poder de transformar sua vida!'];
